+++ Frontend/src/lib/images/index.ts
@@ -3,7 +3,7 @@ export {default as logo} from './logo.svg';
 export {default as logoBeside} from './logo-beside.svg';
 
 export {default as github} from './github.svg';
-export {default as session} from './session.svg';
+export {default as simplex} from './simplex.svg';
 export {default as tor} from './tor.svg';
 
 export {default as info} from './info.svg';